requirements.txt
streamlit==1.11.0
gspread==5.12.0
oauth2client==4.1.3
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
altair==4.2.2